Went through and fixed all instances where Algorithm was referenced instead of AlgorithmSuite.  Changed messages in the docs accordingly. Now if you go to RTD you will see everything links together nicely.
What I did not do: change the variable names from "algorithm" to i.e.
"algorithm_suite".  Mainly because this is a rabbit hole and not immediately
important I think.  To do it you need to change every single thing that
calls another thing that calls another thing that has "algorithm" instead
of "algorithm_suite" and I wanted to keep it simple.

I also found what I think was an incorrect test.  In
/test/integration/test_i_aws_encryption_sdk_client.py in function
test_remove_bad_client() there was an assertion error for my tox env tests
that the dict containing the regional clients was not empty when it should
have been supposedly.  I believe that it was not meant for it to be empty,
but for the bad client "us-fakey-12" to be removed, and it was assumed that
in __init__ for KMSMasterKeyProvider() no regional clients were added, but
this is false because there is a default added depending on some features
of the botocore session etc...  So when it was checking to see if it was
empty it wanted to see that a bad client was added, and then was rightfully
removed -> back to an empty dict, but if the dict did not start out empty
the test would fail.  So I made a one line change of this test to make it
test that specifically the dict did not contain the exact bad client that
the test was using.

Please get back to me on this final change because you guys know this SDK
way better than me and I don't want to break anything.  Thanks!