Explicitly override version when validating release
Otherwise it just pulls down "latest", and with eventual consistency on PyPi
we might end up pulling down a previous version and not actually validating
our release.

This has the downside that, again due to eventual consistency, the new
version might not show up yet causing the build to fail. Since this build
is idempotent this is not the end of the world, but we may want to make
this more robust in the future.