refactor(check_python_version): link directly to support policy
Replaces check_python_version with _warn_deprecated_python.
_warn_deprecated_python is copied from the AWS SDK for Python (boto3/compat.py) and refactored for the ESDK.
No Error will ever be logged, only a warning for Pythons 2.7, 3.4, and 3.5.

Testing is added that not only checks the logic of the function, but also parses the SUPPORT_POLICY.rst for dates and ensures that the dates saved in compatability.py are accurate.