chore(oracle): Upgrade decrypt_oracle to py3.9
AWS Lambda dropped support for py3.6 a while ago.
We have not deployed the oracle in 2 years.
Given our infrequent deployments of the lambda,
the best course of action is deploy the latest python3 supported
by AWS Lambda, 3.9.

I also removed pointless `doc` references in the tox.
The decrypt_oracle does not have docs.
It is not for public comsumption.